Added necessary types to MobileDetector.
