+++ src/MobileDetector/MobileDetector.tsx
@@ -1,6 +1,22 @@
-import React from 'react';
-import MobileDetect from 'mobile-detect';
+import * as React from 'react';
+import * as MobileDetect from 'mobile-detect';
 import { props, t } from '../utils';
+
+export type ChildrenArgs = {
+  isDesktop: boolean,
+  isMobile: boolean,
+  isPhone: boolean,
+  isTablet: boolean
+};
+
+export type MobileDetectorProps = {
+  /** children must be passed as function so to propagte context correctly. Environment info is also passed as first argument to the callback */
+  children: (childrenArgs: ChildrenArgs) => JSX.Element
+  /** ignores real device type and considers it as desktop */
+  forceDesktop?: boolean,
+  /** custom user-agent */
+  userAgent?: string
+}
 
 export const Props = {
   children: t.Function,
@@ -10,12 +26,9 @@ export const Props = {
 
 /**
  * Top-level component which detects device type and passes this info to children as context
- * @param children - children must be passed as function so to propagte context correctly. Environment info is also passed as first argument to the callback
- * @param forceDesktop - ignores real device type and considers it as desktop
- * @param userAgent - custom user-agent
  */
 @props(Props)
-export default class MobileDetector extends React.Component {
+export default class MobileDetector extends React.Component<MobileDetectorProps> {
 
   static childContextTypes = {
     isDesktop: React.PropTypes.bool.isRequired,
@@ -36,11 +49,11 @@ export default class MobileDetector extends React.Component {
     };
   };
 
-  isPhone = (md) => !this.props.forceDesktop && !!md.phone();
+  isPhone = (md: MobileDetect) => !this.props.forceDesktop && !!md.phone();
 
-  isTablet = (md) => !this.props.forceDesktop && !!md.tablet();
+  isTablet = (md: MobileDetect) => !this.props.forceDesktop && !!md.tablet();
 
-  isMobile = (md) => !this.props.forceDesktop && !!md.mobile();
+  isMobile = (md: MobileDetect) => !this.props.forceDesktop && !!md.mobile();
 
   render() {
     return this.props.children(this.getEnvironmentInfo());
