--- src/MobileDetector/index.d.ts
@@ -1,2 +1,0 @@
-import MobileDetector from './MobileDetector';
-export default MobileDetector;
