+++ src/MobileDetector/index.ts
@@ -1,2 +1,3 @@
-import MobileDetector from './MobileDetector';
+import MobileDetector, { MobileDetectorProps } from './MobileDetector';
 export default MobileDetector;
+export { MobileDetectorProps };
